refactor(db): drop top-level re-exports; import from submodules
Keep moutils.db.__init__ docstring-only. Connection/Cursor are internal and
no longer exposed; import connectors from their submodules
(from moutils.db.posthog / moutils.db.datasette import ...).
