+++ src/moutils/db/__init__.py
@@ -1,8 +1,10 @@
 """Read-only DB-API 2.0 connections over REST query APIs, for marimo SQL cells.
 
-Assign a connection to a notebook variable and marimo detects it as a SQL engine:
+Import a connection from its submodule and assign it to a notebook variable;
+marimo then detects it as a SQL engine and you can query it from SQL cells:
 
-    from moutils.db import PostHogConnection, DatasetteConnection
+    from moutils.db.posthog import PostHogConnection
+    from moutils.db.datasette import DatasetteConnection
 
     posthog = PostHogConnection(api_key="phx_...", project_id=123)
     datasette = DatasetteConnection("https://example.com", "mydb")
@@ -10,15 +12,3 @@ Assign a connection to a notebook variable and marimo detects it as a SQL engine
 PostHog uses ``requests`` (a base moutils dependency); Datasette uses ``httpx``,
 available via the optional extra: ``pip install moutils[db]``.
 """
-
-from ._core import Connection, Cursor
-from .datasette import DatasetteConnection, databases
-from .posthog import PostHogConnection
-
-__all__ = [
-    "PostHogConnection",
-    "DatasetteConnection",
-    "databases",
-    "Connection",
-    "Cursor",
-]
